Move download-large-files endpoint to packet controller
